requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic[email]==2.5.0
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
Jinja2==3.1.2
pytest==7.4.0
httpx==0.25.0
black==24.8.0
ruff==0.1.0